Addition of about us page and sorting javascript file
The about us page is created and is responsive.

The sort javascript file works on the shop page that allows the user to
sort the items by price, either high to low or low to high.

Addition and renaming of png's

Slight changes to other html files to fix small bugs.